Move ThemeConfig to shared folder
